move buildOrderHtml to helpers
